DarrenS: Delete opening action, delete opening GUI, improved delete appointment (#195)
* DarrenS: New appointment validation, silence SQL, GUI touch-ups
* Fixed delete opening message not being displayed right
* DarrenS: Fixed all null checking errors in SQL DAOs
* DarrenS: Stop opening creation from saturday from bleeding over to tuesday

+++ src/main/java/edu/lwtech/csd297/teachersfirst/actions/DeleteAppointmentAction.java
@@ -27,16 +27,29 @@ public class DeleteAppointmentAction extends ActionRunner {
 		}
 		final Appointment appointment = DataManager.getAppointmentDAO().retrieveByID(appointmentIdInt);
 		if (appointment == null) {
-			this.SendPostReply("/appointments", "", "Appointment " + appointmentIdInt + " not found!");
+			this.SendPostReply("/appointments", "", "Appointment " + appointmentIdString + " not found!");
 			return;
 		}
+
+		// Make sure the person has the authority
+		if (!isAdmin && !appointment.getIsMyAppointment(uid)) {
+			this.SendPostReply("/appointments", "", "Not your appointment, cannot cancel.");
+			return;
+		}
+
+		// Make sure even if they have the authority, whether they can delete things in the past
+		if (!isAdmin && DateHelpers.isInThePast(appointment.getEndTime().toLocalDateTime())) {
+			this.SendPostReply("/appointments", "", "Appointment has already happened, cannot delete.");
+			return;
+		}
+
 		logger.debug("Attempting to delete appointment " + appointment.toString() + " ...");
 		
 		DataManager.getAppointmentDAO().delete(appointmentIdInt);
-		logger.info(DataManager.getAppointmentDAO().size() + " records total");
-		logger.debug("Deleted appointment ID: [{}]", appointmentIdInt);
+		//logger.info(DataManager.getAppointmentDAO().size() + " records total");
+		logger.debug("Deleted appointment ID: [{}]", appointmentIdString);
 		
-		this.SendPostReply("/appointments", "", "Appointment " + appointmentIdInt + " deleted!");
+		this.SendPostReply("/appointments", "", "Appointment " + appointmentIdString + " deleted!");
 		return;
 	}
